[Decomp][GameObject] Cohesion split: extract Use handlers to GameObjectUse.cpp
